requirements.txt
requests==2.25.1
eth-brownie==1.19.2